+++ .babelrc
@@ -1,3 +1,4 @@
 {
-    "presets": ["@babel/preset-env"]
+    "presets": ["@babel/preset-env"],
+    "plugins": ["@babel/plugin-transform-object-assign"]
 }
