+++ .eslintrc.json
@@ -7,6 +7,9 @@
         "compat"
     ],
     "rules": {
+        "import/no-extraneous-dependencies": ["error", {
+            "devDependencies": true
+        }],
         "compat/compat": "error"
     }
 }
