--- jest.config.json
@@ -1,18 +1,0 @@
-{
-    "preset": "jest-puppeteer",
-    "verbose": true,
-    "collectCoverage": true,
-    "coverageReporters": [
-        "json-summary",
-        "text",
-        "lcov"
-    ],
-    "collectCoverageFrom": [
-        "**/*.{js}",
-        "!**/node_modules/**",
-        "!**/build/**",
-        "!**/sample/**",
-        "!**/jest-coverage-report/**",
-        "!webpack.*.js"
-    ]
-}
