+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@openmind/zero-events",
-  "version": "1.0.1",
+  "version": "1.1.0",
   "description": "The openmind events management",
   "repository": {
     "type": "git",
@@ -11,6 +11,7 @@
   "license": "MIT",
   "devDependencies": {
     "@babel/core": "^7.3.4",
+    "@babel/plugin-transform-object-assign": "^7.2.0",
     "@babel/preset-env": "^7.3.4",
     "babel-loader": "^8.0.5",
     "dmd-bitbucket": "^0.1.10",
@@ -20,7 +21,7 @@
     "eslint-plugin-compat": "^2.7.0",
     "eslint-plugin-import": "^2.16.0",
     "jest": "^24.1.0",
-    "jest-puppeteer": "^4.0.0",
+    "jest-coverage-badges": "^1.1.2",
     "jquery": "^3.3.1",
     "jsdoc-to-markdown": "^4.0.1",
     "semver": "5.5.0",
@@ -32,11 +33,10 @@
   },
   "dependencies": {
     "@babel/polyfill": "^7.2.5",
-    "jest-coverage-badges": "^1.1.2",
     "uuid": "^3.3.2"
   },
   "scripts": {
-    "dev": "webpack-dev-server --config webpack-config.js",
+    "dev": "webpack-dev-server --config webpack-dev.js",
     "build": "webpack --config webpack-prod.js",
     "test": "jest",
     "test:coverage": "npm test -- --coverage",
@@ -45,6 +45,16 @@
     "lint": "node_modules/.bin/eslint ./src"
   },
   "jest": {
+    "verbose": true,
+    "collectCoverage": true,
+    "collectCoverageFrom": [
+      "**/*.{js}",
+      "!**/node_modules/**",
+      "!**/build/**",
+      "!**/sample/**",
+      "!**/jest-coverage-report/**",
+      "!webpack.*.js"
+    ],
     "coverageReporters": [
       "json-summary"
     ]
