+++ webpack-common.js
@@ -1,0 +1,14 @@
+module.exports = {
+  module: {
+    rules: [{
+      test: /\.js$/,
+      exclude: /(node_modules)/,
+      use: {
+        loader: 'babel-loader',
+        options: {
+          presets: ['@babel/preset-env'],
+        },
+      },
+    }],
+  },
+};
